refactor: persistence 계층 클래스명 통일
