+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -2,22 +2,22 @@ package roomescape.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
+import roomescape.dao.ReservationTimeDao;
 import roomescape.domain.reservationtime.ReservationTime;
 import roomescape.dto.reservationtime.ReservationTimeCreateRequest;
 import roomescape.dto.reservationtime.ReservationTimeResponse;
-import roomescape.respository.ReservationTimeRepository;
 
 @Service
 public class ReservationTimeService {
 
-    private final ReservationTimeRepository reservationTimeRepository;
+    private final ReservationTimeDao reservationTimeDao;
 
-    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
-        this.reservationTimeRepository = reservationTimeRepository;
+    public ReservationTimeService(ReservationTimeDao reservationTimeDao) {
+        this.reservationTimeDao = reservationTimeDao;
     }
 
     public List<ReservationTimeResponse> findAll() {
-        List<ReservationTime> allReservationTimes = reservationTimeRepository.findAll();
+        List<ReservationTime> allReservationTimes = reservationTimeDao.readAll();
         return allReservationTimes.stream()
                 .map(ReservationTimeResponse::from)
                 .toList();
@@ -25,14 +25,14 @@ public class ReservationTimeService {
 
     public ReservationTimeResponse add(ReservationTimeCreateRequest request) {
         ReservationTime reservationTime = request.toDomain();
-        ReservationTime result = reservationTimeRepository.add(reservationTime);
+        ReservationTime result = reservationTimeDao.create(reservationTime);
         return ReservationTimeResponse.from(result);
     }
 
     public void delete(Long id) {
         validateNull(id);
         validateNotExist(id);
-        reservationTimeRepository.delete(id);
+        reservationTimeDao.delete(id);
     }
 
     private void validateNull(Long id) {
@@ -42,7 +42,7 @@ public class ReservationTimeService {
     }
 
     private void validateNotExist(Long id) {
-        if (!reservationTimeRepository.exist(id)) {
+        if (!reservationTimeDao.exist(id)) {
             throw new IllegalArgumentException("해당 아이디를 가진 예약 시간이 존재하지 않습니다.");
         }
     }
